+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 Next
 ====
 
+- Allow finer grain for the requests cache (#487)
+
 Version 1.2.27 - 2024-09-10
 ===========================